+++ sdk/cosmos/azure-cosmos-test/CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Release History
 
+### 1.0.0-beta.8 (Unreleased)
+
+#### Features Added
+
+#### Breaking Changes
+
+#### Bugs Fixed
+* Fixed an issue where `FaultInjectionRule` can not apply on partition level when using `Gateway` Mode and non-session consistency - See [40005](https://github.com/Azure/azure-sdk-for-java/pull/40005)
+
 ### 1.0.0-beta.7 (2024-05-03)
 
 #### Bugs Fixed